requirement.txt
Django==2.2.6
django-environ==0.4.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
psycopg2-binary==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
